requirements.txt
absl-py==0.12.0
asgiref==3.2.10
astor==0.8.1
cachetools==4.2.2
certifi==2020.6.20
cffi==1.14.5
chardet==3.0.4
cryptography==3.4.7
defusedxml==0.7.1
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.1
django-allauth==0.44.0
gast==0.2.2
google-auth==1.30.1
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.38.0
gunicorn==20.0.4
h5py==2.10.0
idna==2.8
importlib-metadata==4.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.3.4
numpy==1.20.1
oauthlib==3.1.0
opt-einsum==3.3.0
Pillow==5.4.1
protobuf==3.17.1
psycopg2-binary==2.8.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.1.0
python3-openid==3.2.0
pytz==2020.1
requests==2.21.0
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.4.1
six==1.16.0
sqlparse==0.3.1
static3==0.7.0
tensorboard==2.1.1
tensorflow==2.1.0
tensorflow-estimator==2.1.0
termcolor==1.1.0
typing-extensions==3.10.0.0
urllib3==1.24.1
Werkzeug==2.0.1
whitenoise==5.2.0
wrapt==1.12.1
zipp==3.4.1
